WIP
TODO :
-finir les calcul en fonction de toutes les metamagies
-tout faire niveau conversion
