+++ app/src/main/java/stellarnear/yfa_dealer/Spells/SpellList.java
@@ -17,6 +17,10 @@ public class SpellList extends AppCompatActivity {
         listSpells.add(spell);
     }
 
+    public void addAll(List<Spell> spells) {
+        listSpells.addAll(spells);
+    }
+
     public SpellList filterByRank(int rank){
         SpellList sel_list = new SpellList();
 
@@ -56,4 +60,6 @@ public class SpellList extends AppCompatActivity {
     public List<Spell> asList() {
         return listSpells;
     }
+
+
 }
