+++ app/src/main/res/drawable/ic_play_circle_filled_gradient.xml
@@ -1,8 +1,8 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-        android:width="25dp"
-        android:height="25dp"
-        android:viewportWidth="25.0"
-        android:viewportHeight="25.0"
+        android:width="24dp"
+        android:height="24dp"
+        android:viewportWidth="24.0"
+        android:viewportHeight="24.0"
     xmlns:aapt="http://schemas.android.com/aapt">
     <path
